change trigger, do not translate

+++ main.py
@@ -3,8 +3,10 @@ from typing import Callable
 import os
 import openai
 
+
 load_dotenv()
 openai.api_key = os.getenv("OPENAI_API_KEY")
+
 
 def get_wav_file(file_ogg: str) -> str:
   from pydub import AudioSegment
@@ -13,18 +15,29 @@ def get_wav_file(file_ogg: str) -> str:
   x.export(file_wav, format='wav')
   return file_wav
 
+
 def recognize(file_wav: str) -> str:
-  return openai.Audio.transcribe("whisper-1", open(file_wav, 'rb'), response_format= 'text')
+  response = openai.Audio.transcribe(
+    model="whisper-1",
+    file=open(file_wav, 'rb'),
+    response_format='text',
+  )
+  return response
+
 
 def sanitize(text: str) -> str:
+  prompt = """
+    Remove parasite words from next message, do not change other words.
+    Do not change language.
+    Answer me just with santizied text.
+  """
+
   completion = openai.ChatCompletion.create(
     model="gpt-3.5-turbo",
-    messages=[
-      {'role': 'system',  'content': 'Remove parasite words from next message, do not change other words. Answer me just with santizied text.'},
-      {'role': 'user', 'content': text},
-    ]
+    messages=[{'role': 'system',  'content': prompt}, {'role': 'user', 'content': text}],
   )
   return completion.choices[0].message.content
+
 
 async def handle(file_ogg: str, log: Callable[[str], None]) -> str:
   try:
@@ -39,7 +52,9 @@ async def handle(file_ogg: str, log: Callable[[str], None]) -> str:
 
     return sanitized
   finally:
-    os.remove(file_wav)
+    if os.path.exists(file_wav):
+      os.remove(file_wav)
+
 
 def start_bot():
   from pyrogram import Client, filters
@@ -52,11 +67,16 @@ def start_bot():
       async def log(message: str):
         await response.edit_text("__" + message + "__")
       await response.edit_text(await handle(file_ogg, log))
+    except Exception as e:
+      response and await response.delete()
+      raise e
     finally:
-      os.remove(file_ogg)
+      if os.path.exists(file_ogg):
+        os.remove(file_ogg)
 
   app = Client("text", os.getenv("TELEGRAM_API_ID"), os.getenv("TELEGRAM_API_HASH"))
-  app.add_handler(MessageHandler(handle_voice, filters.voice))
+  app.add_handler(MessageHandler(handle_voice, filters.voice & filters.me))
   app.run()
+
 
 start_bot()
